Fix race condition in MatchProductsArrayAdapter
When adding a new item to a shopping list, the 'product' text field
shows live suggestions. These are updated whenever the input changes.
A reference to the current suggestions list is sent to `publishResults`
inside the `FilterResults` argument, which iterates it to update the
view.
However, it is possible to keep typing and hence modify said list
while it is still being iterated, causing a
`ConcurrentModificationException` inside the for-loop in
`publishResults`.

This commit fixes that by making the suggestions list a local
variable.