Fix/add sorting of quantites, locations and stores
The bottom sheets for selecting quantities, stores and locations
that open up when editing product master data did not contain the
elements in a sorted manner.

This commit fixes that.

However, there is a hacky workaround in here:
The "None selected" entry for both quantities and stores needs to
be on top in order to not be added again and again and again.
The algorithm simply checks the first element for its ID...

Therefore this, the ID = -1 gets pushed to the top during the
sorting *by name*, which is kind of counter-intuitive...