Only disable move-on-open switch if no default consume location selected
The server UI doesn't disable the checkbox. The setting can be useful
even with the default location being the same as the default consume
location: The specific item being opened might not be in the
default consume location already, and the user might want it to be moved
there in that case.