Add a setting to enable or disable the use of the amount.
When a user creates a new product by scanning a barcode,
the amount of the initial entry is also used for barcode generation.

This behavior is useful when the barcode represents a multi-pack.
However, it can be misleading when scanning a single product while
multiple individual units are present.
For example, scanning one bottle of milk while having several
bottles should not imply that the barcode represents multiple units.